Fix Vercel function errors: consolidate chain configs, fix imports, match contract EIP712 structure
