+++ lib/viem.ts
@@ -1,10 +1,8 @@
 import { createWalletClient, http, parseEther } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
-import { base, mainnet, celo, defineChain } from "viem/chains";
-import { weiForUsd } from "./price";
+import { base, celo, defineChain } from "viem/chains";
 
 export const CHAINS = {
-  eth: mainnet,
   base,
   celo,
   mon: defineChain({
@@ -16,20 +14,22 @@ export const CHAINS = {
 } as const;
 
 export const CONTRACTS = {
-  eth: process.env.CONTRACT_ETH as `0x${string}`,
+  base: process.env.CONTRACT_BASE as `0x${string}`,
   celo: process.env.CONTRACT_CELO as `0x${string}`,
   mon: process.env.CONTRACT_MON as `0x${string}`,
+} as const;
+
+export const RPCS = {
+  base: process.env.BASE_RPC!,
+  celo: process.env.CELO_RPC!,
+  mon: process.env.MON_RPC!,
 } as const;
 
 const SIGNER_HEX = (process.env.GAS_SIGNER_PRIVATE_KEY || process.env.SIGNER_PK) as `0x${string}`;
 export const signerAccount = privateKeyToAccount(SIGNER_HEX);
 
 export function signerClient(chain: keyof typeof CHAINS) {
-  const rpc =
-    chain === "eth" ? process.env.ETH_RPC! :
-    chain === "celo" ? process.env.CELO_RPC! :
-    chain === "mon" ? process.env.MON_RPC! :
-    process.env.BASE_RPC!;
+  const rpc = RPCS[chain];
   return createWalletClient({ account: signerAccount, chain: CHAINS[chain], transport: http(rpc) });
 }
 
@@ -37,6 +37,9 @@ export function signerClient(chain: keyof typeof CHAINS) {
 export function fixedPointOne(): bigint {
   return parseEther("0.1");
 }
+
+// Import the price helper
+import { weiForUsd } from "./price";
 
 // Re-export for compatibility
 export const amountWeiForUsd = weiForUsd;
